Removed sub only live streams to features
Removed the retrieval of sub only live stream feeds to features due to it looking as not being possible or extremely patched and a feature that would be used by extremely few or noone. 
Not against implementing it in the future but it will not be included in the 2.0 release.